refactor: global enum 위치 변경
