+++ core/src/main/java/com/picus/core/global/common/enums/ApprovalStatus.java
@@ -1,4 +1,4 @@
-package com.picus.core.domain.post.entity;
+package com.picus.core.global.common.enums;
 
 public enum ApprovalStatus {
     PENDING,
